DigitalAttendance/requirements.txt
# ─────────────────────────────────────────────
# Smart Digital Attendance System
# Install: pip install -r requirements.txt
# ─────────────────────────────────────────────

# Web framework
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9       # required for UploadFile

# Database — async PostgreSQL
asyncpg>=0.29.0
psycopg2-binary>=2.9.9        # only needed if any sync path remains

# Config
python-dotenv>=1.0.0
pydantic>=2.6.0

# Face recognition
insightface>=0.7.3
opencv-python>=4.9.0.80
numpy>=1.26.0

# Utilities
aiofiles>=23.2.1              # async file I/O if needed later
